feat: deixando a aplicação bilhete realizando a consulta do filme por id e salvando no banco como implementação base inicial da pratica

+++ src/services/Cinema.Bilhetes/Cinema.Bilhetes.Domain/Bilhetes/Bilhete.cs
@@ -2,7 +2,7 @@
 {
     public class Bilhete
     {
-        public Bilhete(int filmeId, int checkInId, decimal preco)
+        public Bilhete(long filmeId, int checkInId, decimal preco)
         {
             Id = Guid.NewGuid().ToString();
             FilmeId = filmeId;
@@ -12,7 +12,7 @@
         }
 
         public string Id { get; set; }
-        public int FilmeId { get; set; }
+        public long FilmeId { get; set; }
         public int CheckInId { get; set; }
 
         public decimal Preco { get; set; }
